fix descriptors & ilumination

+++ Engine/Source/ModuleShaderDescriptors.cpp
@@ -36,7 +36,7 @@ bool ModuleShaderDescriptors::init()
 }
 
 
-void ModuleShaderDescriptors::createSRV(ID3D12Resource* resource, UINT8 slot)
+void ModuleShaderDescriptors::createSRV(ID3D12Resource* resource, UINT slot)
 {
 	D3D12_SHADER_RESOURCE_VIEW_DESC srvDesc = {};
 	srvDesc.Shader4ComponentMapping = D3D12_DEFAULT_SHADER_4_COMPONENT_MAPPING;
